+++ apps/meteor/app/livechat/imports/server/rest/rooms.ts
@@ -30,7 +30,7 @@ API.v1.addRoute(
 		async get() {
 			const { offset, count } = await getPaginationItems(this.queryParams);
 			const { sort, fields } = await this.parseJsonQuery();
-			const { agents, departmentId, open, tags, roomName, onhold } = this.queryParams;
+			const { agents, departmentId, open, tags, roomName, onhold, queued } = this.queryParams;
 			const { createdAt, customFields, closedAt } = this.queryParams;
 
 			const createdAtParam = validateDateParams('createdAt', createdAt);
@@ -69,6 +69,7 @@ API.v1.addRoute(
 					tags,
 					customFields: parsedCf,
 					onhold,
+					queued,
 					options: { offset, count, sort, fields },
 				}),
 			);