+++ packages/apps/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rocket.chat/apps
 
+## 0.1.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.10.1
+  - @rocket.chat/model-typings@0.5.1
+  </details>
+
 ## 0.1.0
 
 ### Minor Changes