+++ packages/ui-video-conf/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rocket.chat/ui-video-conf
 
+## 8.0.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@8.0.1
+  - @rocket.chat/ui-avatar@4.0.1
+  </details>
+
 ## 8.0.0
 
 ### Minor Changes